refactor(app): RunSummary, RunningProtocol
