+++ app/src/App/OnDeviceDisplayApp.tsx
@@ -36,8 +36,8 @@ import { ProtocolDetails } from '../pages/ODD/ProtocolDetails'
 import { QuickTransferFlow } from '../organisms/QuickTransferFlow'
 import { QuickTransferDashboard } from '../pages/ODD/QuickTransferDashboard'
 import { QuickTransferDetails } from '../pages/ODD/QuickTransferDetails'
-import { RunningProtocol } from '../pages/RunningProtocol'
-import { RunSummary } from '../pages/RunSummary'
+import { RunningProtocol } from '../pages/ODD/RunningProtocol'
+import { RunSummary } from '../pages/ODD/RunSummary'
 import { UpdateRobot } from '../pages/ODD/UpdateRobot/UpdateRobot'
 import { UpdateRobotDuringOnboarding } from '../pages/ODD/UpdateRobot/UpdateRobotDuringOnboarding'
 import { InstrumentsDashboard } from '../pages/ODD/InstrumentsDashboard'
